libaave: add material absorption sound filtering.
Each surface of the room model is associated with a material.
The materials are stored in a table, referenced by their names,
which contains the acoustic reflection coefficients of each material,
specified for 7 frequency bands: 125, 250, 500, 1000, 2000, 4000, 8000Hz.

The filters are designed by frequency sampling (linear interpolation
on magnitude and linear phase) from the combination of the 7 reflection
coeffients of all surfaces/materials involved in the sound reflection path.

From the designed filter frequency coefficients, the corresponding N filter
coefficients (impulse response) are obtained by inverse discrete Fourier
transform (IDFT of length N).

The calculated filter coefficients are then zero-padded to 4N
to account for the overlap-add of both the HRTF and material filters.

Finally, the zero-padded coefficients are converted to the frequency domain
by a discrete Fourier transform (DFT of length 4N).

The filtering of the sounds is then performed on the frequency domain,
by complex multiplication of the frequency coefficients of the sounds
and the material filters.