libaave: increase spacial resolution of auralisation.
Increase the spacial resolution of the auralised sounds from the
HRTF-sized audio blocks (resolution in the order of 1 metre for
128-sample HRTF at 44100Hz) to the audio sample (resolution of 7.8 mm),
at the expense of one extra DFT for each direct and reflected sound.

To avoid the need for three extra DFT per sound, instead of just one,
the sound is resampled to account for moving listener and/or sources.
The resampler implements a zero-order interpolator (sample and hold).
The resampler is controlled by the current listener-source distances.
The distances, which are usually obtained at a much lower frequency
than the audio sampling frequency, are filtered with a single-pole
recursive low-pass filter (RC filter), where the decay parameter can be
optimally adjusted for the specific position sensors or controls used.