libaave: increase the fade-in/out duration of appearing/disappearing sounds.
The length of an HRTF is not enough fade-in/out duration to apply to
appearing/disappearing sounds. Clicks could be heard on sinusoid test sounds.
This commit separates the fade-in/out of appearing/disappearing sounds
from the HRTF crossfades, and allows increasing the fade-in/out duration
independently. 4096 samples (93ms) seem appropriate for 1kHz sinusoids.